lab 1

+++ lab2/DebugPractice/DebugExercise1.java
@@ -4,9 +4,9 @@
  */
 public class DebugExercise1 {
     public static int divideThenRound(int top, int bottom) {
-        int quotient = top / bottom;
-        int result = Math.round(quotient);
-        return result;
+       double quotient = (double) top / bottom;
+       int result= (int) Math.round(quotient);
+       return result;
     }
 
     public static void main(String[] args) {
